+++ SimpleClient/app/src/main/java/pstl/simpleclient/GameException.java
@@ -1,0 +1,10 @@
+package pstl.simpleclient;
+
+/**
+ * Created by Sylux on 31/01/2017.
+ */
+public class GameException extends Throwable {
+    public GameException(String s) {
+        super(s);
+    }
+}