+++ SimpleClient/app/src/main/java/pstl/simpleclient/ServerException.java
@@ -1,0 +1,10 @@
+package pstl.simpleclient;
+
+/**
+ * Created by Sylux on 31/01/2017.
+ */
+public class ServerException extends Throwable {
+    public ServerException(String s) {
+        super(s);
+    }
+}